2nd course lecture updates.

+++ src/com/belsoft/types/TheMathClass.java
@@ -19,10 +19,10 @@ public class TheMathClass {
         result = (int)Math.floor(1.9);
         System.out.println(result); //1
 
-        result = (int)Math.max(4, 5);
+        result = Math.max(4, 5);
         System.out.println(result); //5
 
-        result = (int)Math.min(4, 5);
+        result = Math.min(4, 5);
         System.out.println(result); //4
 
         double randomNumber = Math.random(); // double between 0.0 and 1.0
@@ -31,7 +31,7 @@ public class TheMathClass {
         randomNumber = Math.random() * 100; // double between 0.0 and 100.0
         System.out.println(randomNumber);
 
-        int randomIntNumber = (int) (Math.random() * 100); // double between 0.0 and 100.0
+        int randomIntNumber = (int) (Math.random() * 100); // int between 0.0 and 100.0
         System.out.println(randomIntNumber);
     }
 }
